requirements.txt
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.4
comtypes==1.1.7
decorator==4.4.2
flake8==3.8.4
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geopy==2.0.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
hurry==1.1
hurry.filesize==0.9
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
more-itertools==8.6.0
MouseInfo==0.1.3

oauthlib==3.1.0
Pillow==8.2.0
pprintpp==0.4.0
protobuf==3.14.0
psutil==5.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pyflakes==2.2.0
PyGetWindow==0.0.9
pyjokes==0.6.0
pylint==2.6.0
PyMsgBox==1.0.9
pyperclip==1.8.1
pypiwin32==223
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3
PyRect==0.1.4
PyScreeze==0.1.26
python-dotenv==0.15.0
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.4
pywhatkit==3.2
pywin32==300
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
ratelim==0.1.6
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
wikipedia==1.4.0
wincertstore==0.2
wolframalpha==4.1.1
wrapt==1.12.1
xml-python==0.3.5
xmltodict==0.12.0
